refactor: Update Navbar component
Refactor the Navbar component to improve the layout and remove unnecessary code. The position of the AppBar has been changed to "sticky" for better user experience. The mobile and desktop menus have been restructured and the Cart icon and User avatar have been removed. This commit improves the overall structure and functionality of the Navbar component.